+++ CHANGELOG.md
@@ -30,6 +30,7 @@ and this project adheres to
 - 🐛(frontend) fix broadcast store sync #1846
 - 🐛(helm) use celery resources instead of backend resources
 - 🐛(helm) reverse liveness and readiness for backend deployment
+- 🐛(y-provider) use CONVERSION_FILE_MAX_SIZE settings #1913
 
 ## [v4.5.0] - 2026-01-28